将dwebTemp变为单精度浮点数

+++ KeShe03WifiOLED.ino
@@ -84,7 +84,7 @@ class wenDuDuiBi
   int zhuangtai = 0;
   void biYiXia(int webTemp, float hereTemp)
   {
-    double dwebTemp = webTemp;
+    float dwebTemp = webTemp;
     if (dwebTemp < hereTemp)
       {
         //dx = "小于";
